Change type for timestampInt

+++ src/graphql/schemas.js
@@ -250,7 +250,7 @@ const typeDefs = gql`
     text: String
     textHtml: String
     timestamp: String
-    timestampInt: Float
+    timestampInt: BigInt
     url: String
     project: Project! @relationship(type: "OWNS", direction: IN)
     member: Member! @relationship(type: "DID", direction: IN)
